requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.2.0
async-timeout==4.0.2
attrs==25.3.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorlog==6.9.0
frozenlist==1.3.1
hikari==2.1.1
hikari-lightbulb==2.3.5.post1
idna==3.3
multidict==6.4.3
python-dotenv==0.21.0
PyYAML==6.0.1
typing_extensions==4.12.2
yarl==1.19.0
Flask==3.1.0